Hardcoded path removed
Remove the hardcoded path and replaced it with direct access to the file in the project. To get the data, one must add an "ExperimentsFiles" folder to his project and put the experiments files there. Then, one must make sure that the files in the folder are noted in the "statistics_files_list.txt" file.